M_PI definition update

+++ omower-constants.h
@@ -7,7 +7,7 @@
 #include <stdint.h>
 
 #ifndef M_PI
-  #define M_PI ((float) 3.14159265358979323846)
+  #define M_PI 3.14159265358979323846f
 #endif
 
 #endif
